@@ -1,6 +1,6 @@
 {
   "name": "genesys-spark-components",
-  "version": "3.89.4",
+  "version": "3.93.4",
   "description": "Common webcomponents",
   "license": "MIT",
   "main": "dist/stencil-wrapper.js",
@@ -32,7 +32,7 @@
     "list-shadow-explicitly-set": "./scripts/list-shadow-explicitly-set.js",
     "predev.public": "npm run predev",
     "prepare": "npm run i18n",
-    "prettier": "prettier --loglevel silent --ignore-unknown --write .",
+    "prettier": "prettier --log-level silent --ignore-unknown --write .",
     "prettier-package-json": "prettier-package-json --write ./package.json",
     "release": "standard-version",
     "stencil": "stencil build --prod",
@@ -60,7 +60,7 @@
   },
   "types": "dist/types/index.d.ts",
   "dependencies": {
-    "@floating-ui/dom": "^1.4.2",
+    "@floating-ui/dom": "^1.4.3",
     "@popperjs/core": "^2.11.8",
     "google-libphonenumber": "^3.2.32",
     "intl-messageformat": "^10.3.5",
@@ -72,8 +72,8 @@
     "vega-lite": "5.5.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.22.1",
-    "@babel/preset-env": "^7.22.4",
+    "@babel/core": "^7.22.8",
+    "@babel/preset-env": "^7.22.7",
     "@stencil/core": "^2.22.3",
     "@stencil/less": "^1.0.0",
     "@stencil/react-output-target": "^0.3.1",
@@ -85,27 +85,26 @@
     "@types/puppeteer": "^5.4.7",
     "@types/resize-observer-browser": "^0.1.7",
     "@types/sortablejs": "^1.15.1",
-    "@typescript-eslint/eslint-plugin": "^5.60.1",
-    "@typescript-eslint/parser": "^5.60.1",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
     "axe-core": "4.4.2",
     "axe-live": "^1.0.1",
-    "babel-loader": "^8.3.0",
     "cross-var": "^1.1.0",
-    "eslint": "^8.41.0",
+    "eslint": "^8.44.0",
     "eslint-config-genesys-spark-components": "file:../../shared-configs/eslint-config-genesys-spark-components",
     "eslint-config-prettier": "^8.8.0",
     "file-loader": "^6.2.0",
     "genesys-spark-tokens": "file:../genesys-spark-tokens",
     "glob": "^10.3.0",
     "handlebars": "^4.7.7",
-    "inquirer": "^8.2.5",
+    "inquirer": "^9.2.7",
     "jest": "^27.5.1",
     "jest-cli": "^27.5.1",
     "jest-config": "^27.5.1",
     "jest-environment-node": "^27.5.1",
     "jest-junit": "^13.2.0",
     "jest-silent-reporter": "^0.5.0",
-    "jsdom": "^20.0.3",
+    "jsdom": "^22.1.0",
     "lint-staged": "^13.2.3",
     "marked": "^4.3.0",
     "mutation-observer": "^1.0.3",
@@ -113,24 +112,22 @@
     "npm-run-all": "^4.1.5",
     "postcss-html": "^1.5.0",
     "postcss-less": "^6.0.0",
-    "prettier": "^2.8.8",
+    "prettier": "^3.0.0",
     "prettier-config-genesys-spark-components": "file:../../shared-configs/prettier-config-genesys-spark-components",
     "prettier-package-json": "^2.8.0",
     "puppeteer": "^14.4.1",
-    "rollup": "^2.79.1",
     "rollup-plugin-copy": "^3.4.0",
-    "rollup-plugin-less": "^1.1.3",
     "semver-sort": "^1.0.0",
     "spritesheet-templates": "^10.5.2",
     "spritesmith": "^3.4.1",
     "standard-version": "^9.5.0",
-    "stylelint": "^15.9.0",
+    "stylelint": "^15.10.1",
     "stylelint-config-genesys-spark-components": "file:../../shared-configs/stylelint-config-genesys-spark-components",
     "stylelint-config-html": "^1.1.0",
-    "stylelint-config-recess-order": "^4.2.0",
-    "stylelint-config-standard": "^33.0.0",
+    "stylelint-config-recess-order": "^4.3.0",
+    "stylelint-config-standard": "^34.0.0",
     "svgo": "^3.0.2",
-    "ts-loader": "^9.4.3",
+    "ts-loader": "^9.4.4",
     "typescript": "^4.9.4",
     "weekstart": "^2.0.0",
     "xmldom": "^0.6.0"
