@@ -12,26 +12,27 @@
     "eslint": "eslint . --ignore-path .gitignore --no-error-on-unmatched-pattern --fix",
     "lint-all": "npm-run-all \"eslint\" \"prettier\" \"prettier-package-json\"",
     "lint-staged": "lint-staged",
-    "prettier": "prettier --loglevel silent --ignore-path .gitignore --ignore-unknown --write .",
+    "prettier": "prettier --log-level silent --ignore-unknown --write .",
     "prettier-package-json": "prettier-package-json --write ./package.json",
     "snapshot": "node build/index.mjs snapshot",
     "test": "node test/index.mjs"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.60.1",
-    "@typescript-eslint/parser": "^5.60.1",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
+    "change-case": "^4.1.2",
     "dir-compare": "^4.0.0",
-    "eslint": "^8.41.0",
+    "eslint": "^8.44.0",
     "eslint-config-genesys-spark-components": "file:../../shared-configs/eslint-config-genesys-spark-components",
     "eslint-config-prettier": "^8.8.0",
     "glob": "^10.3.0",
     "lint-staged": "^13.2.3",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.8.8",
+    "prettier": "^3.0.0",
     "prettier-config-genesys-spark-components": "file:../../shared-configs/prettier-config-genesys-spark-components",
     "prettier-package-json": "^2.8.0",
     "style-dictionary": "^3.8.0",
-    "tempy": "^3.0.0",
+    "tempy": "^3.1.0",
     "token-transformer": "^0.0.33"
   },
   "prettier": "prettier-config-genesys-spark-components"
