@@ -1,6 +1,6 @@
 {
   "name": "genesys-spark-examples",
-  "version": "3.89.4",
+  "version": "3.93.4",
   "license": "MIT",
   "scripts": {
     "prebuild": "npm run clean",
@@ -12,7 +12,7 @@
     "generate-versions-file": "./scripts/generate-versions-file.mjs",
     "lint-all": "npm-run-all \"stylelint\" \"eslint\" \"prettier\" \"prettier-package-json\"",
     "lint-staged": "lint-staged",
-    "prettier": "prettier --loglevel silent --ignore-path .gitignore --ignore-unknown --write .",
+    "prettier": "prettier --log-level silent --ignore-path .gitignore --ignore-unknown --write .",
     "prettier-package-json": "prettier-package-json --write ./package.json",
     "stylelint": "stylelint --ignore-path .gitignore --fix \"**/*.{css,html,less}\"",
     "version-sync": "npm version --no-git-tag-version --allow-same-version"
@@ -20,41 +20,42 @@
   "dependencies": {
     "genesys-spark-components": "file:../../packages/genesys-spark-components",
     "iframe-coordinator": "^4.6.1",
-    "monaco-editor": "^0.29.1",
-    "parse5": "^6.0.1"
+    "monaco-editor": "^0.40.0",
+    "parse5": "^7.1.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.22.1",
-    "@babel/preset-env": "^7.22.4",
+    "@babel/core": "^7.22.8",
+    "@babel/preset-env": "^7.22.7",
     "@stencil/webpack": "0.0.6",
     "@types/copy-webpack-plugin": "^10.1.0",
     "@types/webpack-dev-server": "^4.7.2",
-    "@typescript-eslint/eslint-plugin": "^5.60.1",
-    "@typescript-eslint/parser": "^5.60.1",
-    "copy-webpack-plugin": "^9.1.0",
-    "css-loader": "^5.2.7",
-    "eslint": "^8.41.0",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
+    "copy-webpack-plugin": "^11.0.0",
+    "css-loader": "^6.8.1",
+    "eslint": "^8.44.0",
     "eslint-config-genesys-spark-components": "file:../../shared-configs/eslint-config-genesys-spark-components",
     "eslint-config-prettier": "^8.8.0",
     "file-loader": "^6.2.0",
+    "glob": "^10.3.1",
     "less": "^4.1.3",
-    "less-loader": "^10.2.0",
+    "less-loader": "^11.1.3",
     "lint-staged": "^13.2.3",
-    "monaco-editor-webpack-plugin": "^5.0.0",
+    "monaco-editor-webpack-plugin": "^7.1.0",
     "npm-run-all": "^4.1.5",
     "postcss-html": "^1.5.0",
     "postcss-less": "^6.0.0",
-    "prettier": "^2.8.8",
+    "prettier": "^3.0.0",
     "prettier-config-genesys-spark-components": "file:../../shared-configs/prettier-config-genesys-spark-components",
     "prettier-package-json": "^2.8.0",
     "style-loader": "^3.3.3",
-    "stylelint": "^15.9.0",
+    "stylelint": "^15.10.1",
     "stylelint-config-genesys-spark-components": "file:../../shared-configs/stylelint-config-genesys-spark-components",
     "stylelint-config-html": "^1.1.0",
-    "stylelint-config-recess-order": "^4.2.0",
-    "stylelint-config-standard": "^33.0.0",
-    "webpack": "^5.88.0",
-    "webpack-cli": "^4.10.0",
+    "stylelint-config-recess-order": "^4.3.0",
+    "stylelint-config-standard": "^34.0.0",
+    "webpack": "^5.88.1",
+    "webpack-cli": "^5.1.4",
     "webpack-dev-server": "^4.15.0"
   },
   "prettier": "prettier-config-genesys-spark-components"
