@@ -1,14 +1,14 @@
 #! /usr/bin/env node
 
 const fs = require('fs').promises;
-const { glob } = require('glob');
+const { globSync } = require('glob');
 
 async function listComponentTracking() {
-  const files = glob.sync('src/components/@(beta|stable|legacy)/*/*.tsx');
+  const files = globSync('src/components/@(beta|stable|legacy)/*/*.tsx');
 
   const output = await Promise.all(
     files.map(async file => {
-      const contents = await await fs.readFile(file, 'utf8');
+      const contents = await fs.readFile(file, 'utf8');
 
       const importsTracking = contents.includes(
         'import { trackComponent } from'
