@@ -1,8 +1,8 @@
 #! /usr/bin/env node
 
-const { glob } = require('glob');
+const { globSync } = require('glob');
 
-const files = glob.sync('src/components/**/i18n/*.json');
+const files = globSync('src/components/**/i18n/*.json');
 
 files.map(file => {
   console.log(file);
