@@ -1,10 +1,10 @@
 #! /usr/bin/env node
 
 const fs = require('fs').promises;
-const { glob } = require('glob');
+const { globSync } = require('glob');
 
 async function listShadowExplicitlySet() {
-  const files = glob.sync('src/components/@(beta|stable|legacy)/*/*.tsx', {
+  const files = globSync('src/components/@(beta|stable|legacy)/*/*.tsx', {
     ignore: 'src/components/@(beta|stable|legacy)/*/*.functional.tsx'
   });
 
