@@ -1,7 +1,7 @@
 import { Component, h, JSX, Prop, Element } from '@stencil/core';
 import { trackComponent } from '@utils/tracking/usage';
 
-import { GuxCardAccent } from './gux-card.types';
+import { GuxCardAccent, GuxCardAccentV4 } from './gux-card.types';
 
 /**
  * @slot - Content of card.
@@ -23,10 +23,10 @@ export class GuxCard {
    * Card Accent.
    */
   @Prop()
-  accent: GuxCardAccent = 'outline';
+  accent: GuxCardAccent | GuxCardAccentV4 = 'outline';
 
   componentWillLoad(): void {
-    trackComponent(this.root);
+    trackComponent(this.root, { variant: this.accent });
   }
 
   render(): JSX.Element {
