@@ -439,42 +439,28 @@ export class GuxTabAdvancedList {
   }
 
   handleKeyboardScroll(direction: 'forward' | 'backward'): void {
-    const scrollableSection = this.root.querySelector(
-      '.gux-scrollable-section'
-    );
-    const currentTab =
-      this.root.querySelectorAll('gux-tab-advanced')[this.focused];
-
     if (direction === 'forward') {
       if (this.focused < this.tabTriggers.length - 1) {
-        writeTask(() => {
-          if (this.hasScrollbar) {
-            scrollableSection.scrollBy(currentTab.clientWidth, 0);
-          }
-        });
+        if (this.hasScrollbar) {
+          this.scrollRight();
+        }
         this.focusTab(this.focused + 1);
       } else {
-        writeTask(() => {
-          if (this.hasScrollbar) {
-            scrollableSection.scrollBy(-scrollableSection.scrollWidth, 0);
-          }
-        });
+        if (this.hasScrollbar) {
+          this.scrollToStart();
+        }
         this.focusTab(0);
       }
     } else if (direction === 'backward') {
       if (this.focused > 0) {
-        writeTask(() => {
-          if (this.hasScrollbar) {
-            scrollableSection.scrollBy(-currentTab.clientWidth, 0);
-          }
-        });
+        if (this.hasScrollbar) {
+          this.scrollLeft();
+        }
         this.focusTab(this.focused - 1);
       } else {
-        writeTask(() => {
-          if (this.hasScrollbar) {
-            scrollableSection.scrollBy(scrollableSection.scrollWidth, 0);
-          }
-        });
+        if (this.hasScrollbar) {
+          this.scrollToEnd();
+        }
         this.focusTab(this.tabTriggers.length - 1);
       }
     }
@@ -546,25 +532,35 @@ export class GuxTabAdvancedList {
 
   scrollLeft() {
     writeTask(() => {
-      this.root.querySelector('.gux-scrollable-section').scrollBy(-100, 0);
+      this.root
+        .querySelector('.gux-scrollable-section')
+        .scrollBy(-this.root.querySelector('gux-tab-advanced')?.scrollWidth, 0);
     });
   }
 
   scrollRight() {
     writeTask(() => {
-      this.root.querySelector('.gux-scrollable-section').scrollBy(100, 0);
+      this.root
+        .querySelector('.gux-scrollable-section')
+        .scrollBy(this.root.querySelector('gux-tab-advanced')?.scrollWidth, 0);
     });
   }
 
-  scrollUp() {
+  scrollToStart() {
+    const scrollableSection = this.root.querySelector(
+      '.gux-scrollable-section'
+    );
     writeTask(() => {
-      this.root.querySelector('.gux-scrollable-section').scrollBy(0, -100);
+      scrollableSection?.scrollBy(-scrollableSection?.scrollWidth, 0);
     });
   }
 
-  scrollDown() {
+  scrollToEnd() {
+    const scrollableSection = this.root.querySelector(
+      '.gux-scrollable-section'
+    );
     writeTask(() => {
-      this.root.querySelector('.gux-scrollable-section').scrollBy(0, 100);
+      scrollableSection.scrollBy(scrollableSection?.scrollWidth, 0);
     });
   }
 
@@ -603,9 +599,7 @@ export class GuxTabAdvancedList {
       </span>,
       <div class="gux-tab-container">
         <div class="action-button-container">
-          {this.hasScrollbar
-            ? this.renderScrollButton('scrollLeft')
-            : this.renderScrollButton('scrollUp')}
+          {this.renderScrollButton('scrollLeft')}
         </div>
 
         <div
@@ -619,9 +613,7 @@ export class GuxTabAdvancedList {
         </div>
 
         <div class="action-button-container">
-          {this.hasScrollbar
-            ? this.renderScrollButton('scrollRight')
-            : this.renderScrollButton('scrollDown')}
+          {this.renderScrollButton('scrollRight')}
 
           {this.showNewTabButton ? (
             <AddNewTabButton onClick={() => this.newTab.emit()} />
@@ -661,11 +653,6 @@ export class GuxTabAdvancedList {
       case 'scrollRight':
         this.scrollRight();
         break;
-      case 'scrollUp':
-        this.scrollUp();
-        break;
-      case 'scrollDown':
-        this.scrollDown();
     }
   }
 
@@ -675,10 +662,6 @@ export class GuxTabAdvancedList {
         return 'chevron-left';
       case 'scrollRight':
         return 'chevron-right';
-      case 'scrollUp':
-        return 'chevron-up';
-      case 'scrollDown':
-        return 'chevron-down';
     }
   }
 }
