@@ -1,24 +1,40 @@
 import { temporaryWriteTask } from 'tempy';
 import StyleDictionary from 'style-dictionary';
+import { camelCase } from 'change-case';
 
 export function outputStyles(
   rootFolder,
-  filename,
+  setName,
   styleDictionaryReadableTokens
 ) {
+  const formattedSetName = camelCase(setName);
+
   return temporaryWriteTask(
     JSON.stringify(styleDictionaryReadableTokens),
     tempPath => {
       const styleDictionary = StyleDictionary.extend({
         source: [tempPath],
+        transform: {
+          'name/gse': {
+            type: 'name',
+            transitive: false,
+            matcher: () => true,
+            transformer: (token, options) => {
+              return [options.prefix]
+                .concat(token.path.map(camelCase))
+                .join('-');
+            }
+          }
+        },
         platforms: {
           css: {
             transformGroup: 'css',
-            prefix: 'gse',
+            transforms: ['name/gse'],
+            prefix: `gse-${formattedSetName}`,
             buildPath: `${rootFolder}/css/`,
             files: [
               {
-                destination: `${filename}.css`,
+                destination: `gse-${formattedSetName}.css`,
                 format: 'css/variables'
               }
             ],
@@ -28,25 +44,28 @@ export function outputStyles(
           },
           less: {
             transformGroup: 'less',
-            prefix: 'gse',
+            transforms: ['name/gse'],
+            prefix: `gse-${formattedSetName}`,
             buildPath: `${rootFolder}/less/`,
             files: [
               {
-                destination: `${filename}.less`,
+                destination: `gse-${formattedSetName}.less`,
                 format: 'less/variables'
               }
             ],
             options: {
-              showFileHeader: false
+              showFileHeader: false,
+              outputReferences: false
             }
           },
           scss: {
             transformGroup: 'scss',
-            prefix: 'gse',
+            transforms: ['name/gse'],
+            prefix: `gse-${formattedSetName}`,
             buildPath: `${rootFolder}/scss/`,
             files: [
               {
-                destination: `_${filename}.scss`,
+                destination: `_gse-${formattedSetName}.scss`,
                 format: 'scss/variables'
               }
             ],
@@ -56,11 +75,10 @@ export function outputStyles(
           },
           json: {
             transformGroup: 'js',
-            prefix: 'gse',
             buildPath: `${rootFolder}/json/`,
             files: [
               {
-                destination: `${filename}.json`,
+                destination: `gse-${formattedSetName}.json`,
                 format: 'json/nested'
               }
             ],
