@@ -7,6 +7,8 @@ module.exports = {
   ],
   parser: '@typescript-eslint/parser',
   rules: {
-    '@typescript-eslint/no-inferrable-types': 'off'
+    '@typescript-eslint/no-inferrable-types': 'off',
+    '@typescript-eslint/no-redundant-type-constituents': 'off',
+    '@typescript-eslint/no-unsafe-enum-comparison': 'off'
   }
 };
